+++ R/redcap-next-free-record-name.R
@@ -97,13 +97,15 @@ redcap_next_free_record_name <- function(
         value
       )
 
-    } else { # nocov start
+    } else {
+      # nocov start
       value          <- value_error
       outcome_message  <- sprintf(
         "The REDCap determination of the next free record id failed.  The http status code was %i.  The 'raw_text' returned was '%s'.",
         kernel$status_code,
         kernel$raw_text
       )
+      # nocov end
     }
   } else {
     value          <- value_error
@@ -111,7 +113,7 @@ redcap_next_free_record_name <- function(
       "The REDCap determination of the next free record id failed.  The error message was:\n%s",
       kernel$raw_text
     )
-  }  # nocov stop
+  }
 
   if (verbose)
     message(outcome_message)