+++ R/redcap-users-export.R
@@ -152,6 +152,7 @@ redcap_users_export <- function(
       #   to save RAM.  The content is not really necessary with httr's
       #   status message exposed.
     } else {
+      # nocov start
       kernel$success   <- FALSE # Override the 'success' http status code.
       ds_user          <- data.frame() # Return an empty data.frame
       ds_user_form     <- data.frame() # Return an empty data.frame
@@ -160,6 +161,7 @@ redcap_users_export <- function(
         kernel$status_code,
         kernel$raw_text
       )
+      # nocov end
     }
   } else {
     ds_user          <- data.frame() # Return an empty data.frame