+++ tests/manual/test-stress-test-serial.R
@@ -2,8 +2,11 @@ library(testthat)
 
 context("Stress Test - Serial")
 
-uri <- "https://bbmc.ouhsc.edu/redcap/api/"
-token <- "9A81268476645C4E5F03428B8AC3AA7B" #For `UnitTestPhiFree` account on pid=153.
+# Declare the server & user information
+credential <- REDCapR::retrieve_credential_local(
+  path_credential = system.file("misc/example.credentials", package="REDCapR"),
+  project_id      = 153L
+)
 
 read_count <- 2000L
 file_count <- 200L
@@ -42,7 +45,11 @@ expected_outcome_message <- "5 records and 24 columns were read from REDCap in \
 
 for( i in seq_len(read_count) ) {
   expect_message(
-    returned_object <- redcap_read_oneshot(redcap_uri=uri, token=token, raw_or_label="raw"),
+    returned_object <- redcap_read_oneshot(
+      redcap_uri    = credential$redcap_uri,
+      token         = credential$token,
+      raw_or_label  = "raw"
+    ),
     regexp = expected_outcome_message
   )
 
@@ -110,3 +117,5 @@ for( i in seq_len(file_count) ) {
   expect_more_than(info_actual$atime, expected=start_time, label="The downloaded file's last access time should not precede this function's start time.")
   message(i, ": ", returned_object$elapsed_seconds)
 }
+
+rm(credential)