+++ tests/testthat/test-metadata-read.R
@@ -13,7 +13,7 @@ credential_problem <- REDCapR::retrieve_credential_local(
   project_id      = 1425
 )
 
-test_that("Metadata Smoke Test", {
+test_that("Metadata Read Smoke Test", {
   testthat::skip_on_cran()
   expect_message(
     returned_object <- redcap_metadata_read(redcap_uri=credential$redcap_uri, token=credential$token)
@@ -161,3 +161,7 @@ test_that("Metadata Normal", {
   expect_match(returned_object$outcome_message, regexp=expected_outcome_message, perl=TRUE)
   expect_true(returned_object$success)
 })
+
+rm(credential           )
+rm(credential_super_wide)
+rm(credential_problem   )