+++ tests/testthat/test-read-oneshot.R
@@ -8,7 +8,11 @@ credential <- REDCapR::retrieve_credential_local(
 test_that("smoke test", {
   testthat::skip_on_cran()
   expect_message(
-    returned_object <- redcap_read_oneshot(redcap_uri=credential$redcap_uri, token=credential$token)
+    returned_object <-
+      redcap_read_oneshot(
+        redcap_uri    = credential$redcap_uri,
+        token         = credential$token
+      )
   )
 })
 test_that("default", {
@@ -78,7 +82,10 @@ test_that("default", {
 
   expect_message(
     regexp           = expected_outcome_message,
-    returned_object <- redcap_read_oneshot(redcap_uri=credential$redcap_uri, token=credential$token)
+    returned_object <- redcap_read_oneshot(
+      redcap_uri    = credential$redcap_uri,
+      token         = credential$token
+    )
   )
 
   expect_equivalent(returned_object$data, expected=expected_data_frame, label="The returned data.frame should be correct") # dput(returned_object$data)
@@ -89,7 +96,7 @@ test_that("default", {
   expect_true(returned_object$filter_logic=="", "A filter was not specified.")
   expect_match(returned_object$outcome_message, regexp=expected_outcome_message, perl=TRUE)
   expect_true(returned_object$success)
-  system.file("misc/example.credentials", package="REDCapR")
+  # system.file("misc/example.credentials", package="REDCapR")
 
   # expect_equal_to_reference(returned_object$data, file=system.file("test-data/project-simple/variations/default.rds", package="REDCapR"))
   # expect_equal_to_reference(returned_object$data, file="./test-data/project-simple/variations/default.rds")
@@ -967,3 +974,21 @@ test_that("filter - character", {
   expect_match(returned_object$outcome_message, regexp=expected_outcome_message, perl=TRUE)
   expect_true(returned_object$success)
 })
+
+test_that("bad token -Error", {
+  testthat::skip_on_cran()
+  expected_outcome_message <- "The REDCapR read/export operation was not successful\\."
+
+  expect_message(
+    returned_object <-
+      redcap_read_oneshot(
+        redcap_uri    = credential$redcap_uri,
+        token         = "BAD00000000000000000000000000000"
+      ),
+    expected_outcome_message
+  )
+  testthat::expect_false(returned_object$success)
+  testthat::expect_equal(returned_object$status_code, 403L)
+  testthat::expect_equal(returned_object$raw_text, "ERROR: You do not have permissions to use the API")
+})
+rm(credential)