+++ NEWS.md
@@ -2,8 +2,7 @@ Upcoming Versions
 ==========================================================
 
 In the future:
-* `redcap_read()` and `redcap_read_oneshot()` allows caller to specify data types for columns.
-
+...
 
 Version 0.11 (Released ?)
 ==========================================================
@@ -20,11 +19,17 @@ Version 0.11 (Released ?)
 
 * [`reader::type_convert()`](https://readr.tidyverse.org/reference/type_convert.html) is used *after* all the batches are stacked on top of each other.  This way, batches cannot have incompatible data types as they're combined. (#257; thanks @isaactpetersen #245)  Consequently, the `guess_max` parameter in `redcap_read()` no longer serves a purpose, and has been soft-deprecated. (#267)
 
+* [`redcap_metadata_write()`](https://readr.tidyverse.org/reference/redcap_metadata_write.html) writes to the project's metadata. (#274, @felixetorres) 
+
 ### Stability Features
 
 * `httr::content()` (which is inside `kernel_api()`) now processes the returned value as "text/csv", by default.  This should prevent strange characters from tricking the process as the internal variable `raw_text` is being formed. See the [httr::content()`](https://httr.r-lib.org/reference/content.html) documentation for a list of possible values for the `content_type` parameter.  (Thanks to great debugging by @vortexing, #269)
 
 * Similarly, `kernel_api()` now has an `encoding` parameter, which defaults to "UTF-8".  (#270)
+
+### Minor Enhancements
+
+* check for bad field names passed to the read records functions (#288)
 
 ### Corrections